# backend/cmd/api/openapi.yaml
openapi: 3.1.0
info:
    title: University Room Booking API
    description: |
        REST API for the University Room Booking System.
        Allows users to search rooms, view availability timelines, create booking requests,
        and track their status. Administrators approve/reject pending bookings via a dashboard.

        ## Authentication
        - **Access Token**: Short-lived JWT (15 min) stored in an `httpOnly` cookie.
          `Authorization: Bearer <token>` remains supported for API clients.
        - **Refresh Token**: Long-lived token (30 days) stored in an `httpOnly` cookie with rotation on each use.

        ## Pagination
        All list endpoints use cursor-based pagination. The `nextCursor` value from a response
        should be passed as the `cursor` query parameter in the next request.

        ## Time Slots
        All booking times must be in 5-minute increments.
        Working day: 08:00–22:00.
    version: 1.0.0
    contact:
        name: Backend Architecture Team
servers:
    - url: http://localhost:3000/api
      description: Local development
    - url: https://booking.university.edu/api
      description: Production
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: Compatibility access token auth for API clients
        AccessCookieAuth:
            type: apiKey
            in: cookie
            name: accessToken
            description: httpOnly access token cookie
        RefreshCookieAuth:
            type: apiKey
            in: cookie
            name: refreshToken
            description: httpOnly refresh token cookie

    # ────────────────────────────────────────────
    # Reusable Schemas
    # ────────────────────────────────────────────
...
paths:
...
    /auth/register:
        post:
            operationId: register
            tags: [Auth]
            summary: Register a new user
            description: Creates a new user account with `user` role. Returns access token in body for compatibility and sets access/refresh httpOnly cookies.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/RegisterRequest"
                        example:
                            email: john.doe@university.edu
                            password: SecureP@ss123
                            firstName: John
                            lastName: Doe
                            department: Faculty of Computer Science
            responses:
                "201":
                    description: User created successfully
                    headers:
                        Set-Cookie:
                            schema:
                                type: string
                            description: "accessToken=<jwt>; HttpOnly; Secure; SameSite=Strict; Path=/api; Max-Age=900 and refreshToken=<token>; HttpOnly; Secure; SameSite=Strict; Path=/api/auth; Max-Age=2592000"
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        type: object
                                        required: [user, accessToken]
                                        properties:
                                            user:
                                                $ref: "#/components/schemas/User"
                                            accessToken:
                                                type: string
                "400":
                    $ref: "#/components/responses/ValidationError"
                "409":
                    description: Email already registered
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                            example:
                                error:
                                    code: EMAIL_ALREADY_EXISTS
                                    message: This email is already registered
                "429":
                    $ref: "#/components/responses/RateLimited"
    /auth/login:
        post:
            operationId: login
            tags: [Auth]
            summary: Log in
            description: Authenticates with email/password. Returns access token in body for compatibility and sets access/refresh httpOnly cookies.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/LoginRequest"
                        example:
                            email: john.doe@university.edu
                            password: SecureP@ss123
            responses:
                "200":
                    description: Login successful
                    headers:
                        Set-Cookie:
                            schema:
                                type: string
                            description: "accessToken=<jwt>; HttpOnly; Secure; SameSite=Strict; Path=/api; Max-Age=900 and refreshToken=<token>; HttpOnly; Secure; SameSite=Strict; Path=/api/auth; Max-Age=2592000"
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        type: object
                                        required: [user, accessToken]
                                        properties:
                                            user:
                                                $ref: "#/components/schemas/User"
                                            accessToken:
                                                type: string
                "400":
                    $ref: "#/components/responses/ValidationError"
                "401":
                    description: Invalid email or password
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                            example:
                                error:
                                    code: UNAUTHORIZED
                                    message: Invalid email or password
                "429":
                    $ref: "#/components/responses/RateLimited"
    /auth/refresh:
        post:
            operationId: refreshToken
            tags: [Auth]
            summary: Refresh access token
            description: |
                Uses the refresh token from the `refreshToken` httpOnly cookie.
                Performs **token rotation**: the old refresh token is revoked and a new one is issued.
                Also sets a fresh `accessToken` httpOnly cookie.
            security:
                - RefreshCookieAuth: []
            responses:
                "200":
                    description: New tokens issued
                    headers:
                        Set-Cookie:
                            schema:
                                type: string
                            description: New accessToken and refreshToken cookies
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        type: object
                                        required: [accessToken]
                                        properties:
                                            accessToken:
                                                type: string
                "401":
                    description: Refresh token missing, expired, or revoked
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                            example:
                                error:
                                    code: REFRESH_TOKEN_INVALID
                                    message: Refresh token is invalid, expired, or revoked
                "429":
                    $ref: "#/components/responses/RateLimited"
...
    /auth/me:
        get:
            operationId: getMe
            tags: [Auth]
            summary: Get current user profile
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            responses:
                "200":
                    description: Current user data
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        $ref: "#/components/schemas/User"
                "401":
                    $ref: "#/components/responses/Unauthorized"
...
    /buildings:
        get:
            operationId: listBuildings
            tags: [Rooms]
            summary: List active buildings
            description: Returns localized building labels using `X-Locale` or `Accept-Language`.
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            responses:
                "200":
                    description: Building catalog
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/BuildingOption"
    /booking-purposes:
        get:
            operationId: listBookingPurposes
            tags: [Bookings]
            summary: List active booking purposes
            description: Returns localized booking purpose labels using `X-Locale` or `Accept-Language`.
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            responses:
                "200":
                    description: Booking purpose catalog
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/BookingPurposeOption"

    # ── Rooms ───────────────────────────────────
    /rooms:
        get:
            operationId: searchRooms
            tags: [Rooms]
            summary: Search rooms with availability
            description: |
                Returns a paginated list of active rooms with computed availability for the given date.
                Supports filtering by name, time range, equipment, and capacity.
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/DateParam"
                - $ref: "#/components/parameters/SearchParam"
                - name: building
                  in: query
                  schema:
                      type: string
                      default: aviamotornaya
                  description: Building code. Defaults to aviamotornaya.
                - name: timeFrom
                  in: query
                  schema:
                      type: string
                      pattern: "^([01]\\d|2[0-3]):([0-5][05])$"
                  description: Start of desired time range (HH:mm, 5-minute granularity)
                - name: timeTo
                  in: query
                  schema:
                      type: string
                      pattern: "^([01]\\d|2[0-3]):([0-5][05])$"
                  description: End of desired time range
                - name: equipment
                  in: query
                  schema:
                      type: string
                  description: "Comma-separated equipment UUIDs. Rooms must have ALL listed equipment."
                  example: "uuid1,uuid2"
                - name: minCapacity
                  in: query
                  schema:
                      type: integer
                      minimum: 1
                  description: Minimum room capacity
                - $ref: "#/components/parameters/LimitParam"
                - $ref: "#/components/parameters/CursorParam"
            responses:
                "200":
                    description: Paginated room list
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data, meta]
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/RoomCard"
                                    meta:
                                        $ref: "#/components/schemas/CursorPaginationMeta"
                "400":
                    description: Invalid time range
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                "401":
                    $ref: "#/components/responses/Unauthorized"

        post:
            operationId: createRoom
            tags: [Rooms]
            summary: Create a room (admin)
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateRoomRequest"
                    multipart/form-data:
                        schema:
                            $ref: "#/components/schemas/CreateRoomMultipartRequest"
            responses:
                "201":
                    description: Room created
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        $ref: "#/components/schemas/RoomFull"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "403":
                    $ref: "#/components/responses/Forbidden"
    /rooms/{roomId}:
        get:
            operationId: getRoomDetail
            tags: [Rooms]
            summary: Room detail with time-slot timeline
            description: |
                Returns full room info and a timeline of time slots for the given date.
                Slot statuses: `available`, `occupied` (confirmed by another user),
                `pending` (pending by another user), `yours` (current user's booking of any status).
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/RoomIdParam"
                - $ref: "#/components/parameters/DateParam"
            responses:
                "200":
                    description: Room detail with timeline
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        $ref: "#/components/schemas/RoomDetail"
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "404":
                    description: Room not found or inactive
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorResponse"
                            example:
                                error:
                                    code: ROOM_NOT_FOUND
                                    message: Room not found

        put:
            operationId: updateRoom
            tags: [Rooms]
            summary: Update a room (admin)
            description: Full replacement of room data.
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/RoomIdParam"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateRoomRequest"
                    multipart/form-data:
                        schema:
                            $ref: "#/components/schemas/UpdateRoomMultipartRequest"
            responses:
                "200":
                    description: Room updated
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        $ref: "#/components/schemas/RoomFull"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "403":
                    $ref: "#/components/responses/Forbidden"
                "404":
                    $ref: "#/components/responses/NotFound"

        delete:
            operationId: deleteRoom
            tags: [Rooms]
            summary: Soft-delete a room (admin)
            description: "Sets `is_active = false`. Existing bookings are NOT cancelled."
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/RoomIdParam"
            responses:
                "204":
                    description: Room deactivated
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "403":
                    $ref: "#/components/responses/Forbidden"
                "404":
                    $ref: "#/components/responses/NotFound"

    # ── Equipment ───────────────────────────────
    /equipment:
        get:
            operationId: listEquipment
            tags: [Equipment]
            summary: List all equipment types
            description: Returns the full equipment catalogue. Used by the frontend for filter rendering.
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            responses:
                "200":
                    description: Equipment list
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/EquipmentItem"
                            example:
                                data:
                                    - id: eq-uuid-1
                                      name: Projector
                                      icon: IconVideo
                                    - id: eq-uuid-2
                                      name: Whiteboard
                                      icon: IconPresentation
                                    - id: eq-uuid-3
                                      name: Computers
                                      icon: IconDeviceDesktop
                                    - id: eq-uuid-4
                                      name: Microphone
                                      icon: IconMicrophone
                                    - id: eq-uuid-5
                                      name: Live Stream
                                      icon: IconBroadcast
                                    - id: eq-uuid-6
                                      name: Spatial Audio
                                      icon: IconVolume
                                    - id: eq-uuid-7
                                      name: Wi-Fi
                                      icon: IconWifi
                                    - id: eq-uuid-8
                                      name: Linux Nodes
                                      icon: IconTerminal2
                "401":
                    $ref: "#/components/responses/Unauthorized"

    # ── Bookings ────────────────────────────────
...
    /bookings/my:
        get:
            operationId: getMyBookings
            tags: [Bookings]
            summary: My active bookings
            description: |
                Returns the current user's bookings that are still upcoming
                (`pending` or `confirmed`, and `booking_date + end_time` is in the future).
                Sorted ascending by date/time.
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/SearchParam"
                - $ref: "#/components/parameters/LimitParam"
                - $ref: "#/components/parameters/CursorParam"
            responses:
                "200":
                    description: Paginated active bookings
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data, meta]
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/MyBooking"
                                    meta:
                                        $ref: "#/components/schemas/CursorPaginationMeta"
                "401":
                    $ref: "#/components/responses/Unauthorized"
    /bookings/my/history:
        get:
            operationId: getMyBookingHistory
            tags: [Bookings]
            summary: My past / inactive bookings
            description: |
                Returns the current user's bookings that are expired, rejected, or cancelled.
                Sorted descending by date (newest first).
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/SearchParam"
                - $ref: "#/components/parameters/LimitParam"
                - $ref: "#/components/parameters/CursorParam"
            responses:
                "200":
                    description: Paginated history bookings
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data, meta]
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/MyBooking"
                                    meta:
                                        $ref: "#/components/schemas/CursorPaginationMeta"
                "401":
                    $ref: "#/components/responses/Unauthorized"
...
    /admin/booking-purposes/{code}:
        put:
            operationId: updateBookingPurpose
            tags: [Admin]
            summary: Update booking purpose labels/status
            parameters:
                - name: code
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Updated booking purpose
        delete:
            operationId: deactivateBookingPurpose
            tags: [Admin]
            summary: Deactivate booking purpose
            parameters:
                - name: code
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "204":
                    description: Deactivated
    /admin/booking-purposes/{code}/reactivate:
        patch:
            operationId: reactivateBookingPurpose
            tags: [Admin]
            summary: Reactivate booking purpose
            parameters:
                - name: code
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Reactivated booking purpose
...